Focused on World Map
NodeWorld is pretty much done. Just need to add in JPanel/JFrame things
and names for image files

WorldMap's constructor and fields are pretty much done. Currently
working on methods to limit player movement, etc.

PlayerWorld.java is a new file. It's the character piece on the World
Map

dumpedMethods.java is a new file. It's an archive for methods that we no
longer need, but still to be admired