Movement System
Movement system. In character, it now takes in current tile and change
in tile, finds the new tile, checks if it is ok, then changes
currentTile to this new one, and changes CharacterID for the new and old
tiles. A potential Issue is that I rely on two different references to
the same DungeonTile. I don't know if that is allowed. In player, I if
bash to get the correct movement. Also eclipse is messing up my spacing.
I slightly improved the DungeonTile as well, by removing an unneeded
constructor parameter.