MaxAreaPolygon Works
MaxAreaPolygon is fixed. The problem was that the definition of
insideness for java.awt.Polygon did not include vertices. We checked for
insideness of points before removing them from the viableList, which
meant that when a new vertex was added, it would never be removed from
the viableList. The program would only continue on if the viableList was
empty, resulting in infinite loops.

The next problem that came up is the 3 point problem - when there are
only 3 starting vertices (ie: a vertex with BOTH largest Y AND X). I
resolved this with if statements that prevent the running of an empty
slope