Many minor changes and fixes
Many systems prepared for future, aka. hoping LWJGL will solve our
problems. I think the 2 step freezing error is related to
Keyboardinput's polling.

*Added is a factor that checks how well the player did that floor. If
they took a lot of damage and didn't manage to collect a lot of money,
then more money spawns next floor. Useful for finetuning difficulty
later.

*Added overloaded constructor for DungeonBuilder so that new floors use
the previous floor's data. Also made some other improvements to new
floor creation such as a changeSpawnRate method.

*Moved the entire Jam class to RandomJam. However, all Jams extend the
abstract Jam class so that Jam specific stuff can be done, such as the
jam-eater

*Fixed revive as a player skill. Interestingly, if another character is
on top of the dead character, it is not overridden. On the next turn it
separates.

In the next update, I just need to finish the AI for the ghost (reviver)
to live