Fixing projectile.
8/24/14:
Now the actual projectile itself works, need to work on animations.

CHANGES:
Character.java:
*Added isMoving boolean that will activate when this character is moving
to allow for animations.

DungeonMain.java:
*Renamed image files to be loaded so that jar creation works. The
capitalization doesn't seem to matter to github though.

*Redid the way projectiles act. Now they use a regular for loop and I
decrement i if a projectile is destroyed.

*Started work on animateMovement() which should allow for the smooth
animation from one point to another in an alotted time.

*In nextLevel(), I reset some important arrayLists to prevent crashes
trying to access that data on different floors.

*Added private coordinate class for use in animation. It just consists
of an x and y value.

Projectile.java:
*Made projectile store only the x and y data from the input tile. Only
that x and y can change.

*Added clause preventing friendly fire by checking if the character on
the tile is of the same faction.

*Made the projectile return true if it was destroyed and false if it
will last to the next turn.

Skills.java:
*Made it so that the projectile will start on the same tile as the
sourceCharacter.