Restructuring of skills to act as an object.
Now skills are in an array and ready to be interchanged :>

CHANGES:
Character.java:
*No longer extends skills (finally)

Player.java:
*Now has an array called hotbarList that keeps track of the skills as an
object in that list. Adding to list is simple, just set up the
constructor so it takes everything needed outside of DungeonMain and
sourceCharacter. To access, simply access that index and use
.run(DungeonMain, sourceCharacter)

*Replaced players skills so that it now uses this system. Depending on
stats though, skills will need to be refreshed.

Skills.java:
*Placed each skill as their own class in this file. They all have a
.run() method and a constructor for other needed values such as range.

Teleball.java:
*Fixed with new skill system.